+++ soya/HISTORY.md
@@ -1,5 +1,17 @@
 # Version 0.0.x
 
+## 0.0.66
+
+- Updated logic of handleChange(), fixed bug recursive handleChange() bug. The
+  new implementation should no longer have inconsistency between redux state
+  and view state.
+- Added simple test case for recursive handleChange().
+
+## 0.0.65
+
+- Add Form.enableField(), Form.enableFields() helper.
+- Add Form.disableField(), Form.disableFields() helper.
+
 ## 0.0.64
 
 - Fix isEqualShallowArray helper bug.